pyproject.toml
[project]
name = "phoenix-llama-index-pydantic-error"
version = "0.1.0"
description = ""
authors = [{ name = "Jeremy-Hibiki", email = "JeremyJiang430@outlook.com" }]
dependencies = [
    # instrument
    "arize-phoenix-otel==0.4.1",
    "openinference-instrumentation-llama-index==3.0.0",
    # llama-index
    "llama-index-core==0.11.3",
    "llama-index-llms-openai==0.2.0",
    "llama-index-llms-openai-like==0.2.0",
    "openai==1.43.0",
    "pydantic==2.8.2",
    # cli
    "rich>=13.8.0",
    "typer>=0.12.5",
]
requires-python = ">= 3.10"
license = { text = "MIT" }

[tool.rye]
managed = true
virtual = true
universal = true
dev-dependencies = ["mypy[tomli]==1.11.2", "ruff==0.6.3"]

[tool.ruff]
line-length = 119
indent-width = 4
target-version = "py310"
builtins = ["typing_extensions"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "FA", "SIM", "I", "UP"]


[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true
